+++ src/Grpc.AspNetCore/GrpcRouteBuilderExtensions.cs
@@ -1,9 +1,28 @@
-﻿using System;
-using Google.Protobuf.Reflection;
+﻿#region Copyright notice and license
+
+// Copyright 2015 gRPC authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#endregion
+
+using System;
+using Grpc.AspNetCore.Internal;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
+using ProtobufServiceDescriptor = Google.Protobuf.Reflection.ServiceDescriptor;
 
-namespace Grpc.AspNetCore
+namespace Microsoft.Extensions.DependencyInjection
 {
     public static class GrpcRouteBuilderExtensions
     {
@@ -28,7 +47,7 @@ namespace Grpc.AspNetCore
             var declaringType = baseType.DeclaringType;
 
             // Get the descriptor
-            var descriptor = declaringType.GetProperty("Descriptor").GetValue(null) as ServiceDescriptor ?? throw new InvalidOperationException("Cannot retrive service descriptor");
+            var descriptor = declaringType.GetProperty("Descriptor").GetValue(null) as ProtobufServiceDescriptor ?? throw new InvalidOperationException("Cannot retrive service descriptor");
 
             foreach (var method in descriptor.Methods)
             {
