--- src/Grpc.AspNetCore/IGrpcServiceActivator.cs
@@ -1,8 +1,0 @@
-﻿namespace Grpc.AspNetCore
-{
-    interface IGrpcServiceActivator<TGrpcService> where TGrpcService : class
-    {
-        TGrpcService Create();
-        void Release(TGrpcService service);
-    }
-}
