--- src/Grpc.AspNetCore/IServerCallHandler.cs
@@ -1,11 +1,0 @@
-﻿
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-
-namespace Grpc.AspNetCore
-{
-    interface IServerCallHandler
-    {
-        Task HandleCallAsync(HttpContext httpContext);
-    }
-}
