+++ test/Grpc.AspNetCore.Tests/DefaultGrpcServiceActivatorTests.cs
@@ -1,6 +1,25 @@
+#region Copyright notice and license
+
+// Copyright 2015 gRPC authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#endregion
+
 using NUnit.Framework;
 using Moq;
 using System;
+using Grpc.AspNetCore.Internal;
 
 namespace Grpc.AspNetCore.Tests
 {
@@ -9,14 +28,14 @@ namespace Grpc.AspNetCore.Tests
         public class GrpcService { }
 
         [Test]
-        public void HubCreatedIfNotResolvedFromServiceProvider()
+        public void GrpcServiceCreatedIfNotResolvedFromServiceProvider()
         {
             Assert.NotNull(
                 new DefaultGrpcServiceActivator<GrpcService>(Mock.Of<IServiceProvider>()).Create());
         }
 
         [Test]
-        public void HubCanBeResolvedFromServiceProvider()
+        public void GrpcServiceCanBeResolvedFromServiceProvider()
         {
             var service = Mock.Of<GrpcService>();
             var mockServiceProvider = new Mock<IServiceProvider>();
